backend/requirements.txt
# Web API
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
pydantic>=2.10,<3.0
pydantic-settings>=2.8,<3.0
python-multipart>=0.0.20,<1.0
sse-starlette>=2.2,<3.0
loguru>=0.7,<1.0
httpx>=0.28,<1.0

# CV / ML
numpy>=2.2,<3.0
pillow>=11.0,<12.0
opencv-python-headless>=4.11,<5.0
albumentations>=2.0,<3.0
rembg>=2.0,<3.0
timm>=1.0,<2.0
torch>=2.6,<3.0
torchvision>=0.21,<1.0
pytorch-lightning>=2.5,<3.0
grad-cam>=1.5,<2.0
scikit-learn>=1.6,<2.0

# RAG
chromadb>=1.0,<2.0
langchain>=0.3,<0.4
langchain-community>=0.3,<0.4
langgraph>=0.4,<0.5
sentence-transformers>=4.0,<5.0
pypdf>=5.4,<6.0

# TTS / translation support
transformers>=4.51,<5.0
accelerate>=1.6,<2.0
scipy>=1.15,<2.0
gTTS>=2.5,<3.0

# Experiment tracking
mlflow>=2.22,<3.0

# Testing
pytest>=8.0,<9.0
